efb/ changes ...

+++ efb/ebk_gui.e
@@ -22,20 +22,28 @@ feature -- Attributes
 
 	ebk_window: EV_TITLED_WINDOW
 
+	ui_thread: EBK_UI_THREAD
+
 feature -- Creation
 
 	make (a_ui_thread: EBK_UI_THREAD)
+		local
+			w: like ebk_window
+			l_screen: EV_SCREEN
 		do
+			ui_thread := a_ui_thread
 			create ev_app
 			create ebk_window
 			ebk_window.set_minimum_size (100, 100)
+			w := ebk_window
+			create l_screen
+				-- Centre the window
+			ebk_window.set_position (
+				(l_screen.width // 2) - (w.width // 2),
+				(l_screen.height // 2) - (w.height // 2))
 			add_to_main_window
 			ebk_window.show
 			ev_app.launch
-
-
-			print ("ebk_window.is_destroyed: " + ebk_window.is_destroyed.out + "%N")
-			print ("Return from ev_app.launch ... %N")
 		end
 
 	add_to_main_window
@@ -77,11 +85,19 @@ feature -- Creation
 		end
 
 	req_close_window
+		local
+			question_dialog: EV_CONFIRMATION_DIALOG
 		do
-			ebk_window.destroy
-			ev_app.process_events
-			ev_app.destroy
-			print ("Close GUI requested ...%N")
+			create question_dialog.make_with_text ("Close the GUI ?")
+			question_dialog.show_modal_to_window (ebk_window)
+
+			if attached question_dialog.selected_button as sb
+			and then sb.is_equal ((create {EV_DIALOG_CONSTANTS}).ev_ok) then
+					-- Destroy the window
+				ebk_window.destroy
+				ev_app.process_events
+				ev_app.destroy
+			end
 		end
 
 feature -- Run tests
@@ -128,6 +144,7 @@ feature -- Run tests
 		local
 			l_test: DIR_COMMS_TEST
 		do
-			create l_test.make; -- l_test.launch
+			create l_test.make (ui_thread); -- l_test.launch
+
 		end
 end
